backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "juno-backend"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",

    # Validation & config
    "pydantic>=2.11",
    "pydantic-settings>=2.7",
    "python-dotenv>=1.0",

    # AWS (Nova Sonic + Bedrock) — smithy-generated async SDK for bidirectional streaming
    "aws-sdk-bedrock-runtime",
    "smithy-aws-core",

    # Anthropic Claude
    "anthropic>=0.52",

    # Firebase
    "firebase-admin>=6.7",

    # Google Calendar
    "google-api-python-client>=2.168",
    "google-auth>=2.40",
    "google-auth-httplib2>=0.2",

    # Google Gen AI SDK (Gemini via Vertex AI backend — nutrition VLM)
    "google-genai>=1.0",

    # Google Cloud Tasks (durable delayed job scheduling)
    "google-cloud-tasks>=2.16",

    # Lambda adapter (for REST handlers deployed to Lambda)
    "mangum>=0.19",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9",
    "pyright>=1.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
